printf format/argument agreement in games/robots/score.c.

+++ games/robots/score.c
@@ -176,7 +176,7 @@ score(score_wfd)
 		if (!done_show && scp->s_uid == uid && scp->s_score == Score)
 			standout();
 		printw("%5.5d %5.5d %-8.8s %-9.9s %5.5d",
-		    (scp - Top) + 1, scp->s_score, scp->s_name,
+		    (int)(scp - Top) + 1, scp->s_score, scp->s_name,
 		    scp->s_auto ? "(autobot)" : "", scp->s_level);
 		if (!done_show && scp->s_uid == uid && scp->s_score == Score) {
 			standend();
